src/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
idna==2.10
mongoengine==0.20.0
multidict==4.7.6
PyJWT==1.7.1
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
twilio==6.44.1
typing-extensions==3.7.4.2
urllib3==1.25.10
websockets==8.1
yarl==1.5.0
